+++ src/test/groovy/ru/vyarus/gradle/plugin/lib/UpstreamKitTest.groovy
@@ -2,11 +2,15 @@ package ru.vyarus.gradle.plugin.lib
 
 import org.gradle.testkit.runner.TaskOutcome
 
+import java.util.zip.ZipFile
+
 /**
- * @author Vyacheslav Rusakov 
- * @since 10.11.2015
+ * @author Vyacheslav Rusakov
+ * @since 13.11.2019
  */
-class JavaLibPluginKitTest extends AbstractKitTest {
+class UpstreamKitTest extends AbstractKitTest {
+
+    String GRADLE_VERSION = '6.0'
 
     def "Check install task"() {
         setup:
@@ -22,7 +26,7 @@ class JavaLibPluginKitTest extends AbstractKitTest {
         """
 
         when: "run pom task"
-        def result = run('install')
+        def result = runVer(GRADLE_VERSION, 'install')
 
 
         String artifactId = projectName()
@@ -35,7 +39,7 @@ class JavaLibPluginKitTest extends AbstractKitTest {
         then: "artifacts deployed"
         deploy.exists()
         def baseName = artifactId + '-1.0'
-        deploy.list() as Set ==
+        withoutModuleFile(deploy) ==
                 ["${baseName}.jar", "${baseName}.pom", "${baseName}-sources.jar", "${baseName}-javadoc.jar"] as Set<String>
     }
 
@@ -53,7 +57,7 @@ class JavaLibPluginKitTest extends AbstractKitTest {
         """
 
         when: "run pom task"
-        def result = run('install')
+        def result = runVer(GRADLE_VERSION, 'install')
 
 
         String artifactId = projectName()
@@ -67,7 +71,7 @@ class JavaLibPluginKitTest extends AbstractKitTest {
         deploy.exists()
         def baseName = artifactId + '-1.0'
         // javadoc will be produced instead of groovydoc! important for maven central
-        deploy.list() as Set ==
+        withoutModuleFile(deploy) ==
                 ["${baseName}.jar", "${baseName}.pom", "${baseName}-sources.jar", "${baseName}-javadoc.jar"] as Set<String>
     }
 
@@ -86,7 +90,7 @@ class JavaLibPluginKitTest extends AbstractKitTest {
         """
 
         when: "run pom task"
-        def result = run('install')
+        def result = runVer(GRADLE_VERSION, 'install')
 
 
         String artifactId = projectName()
@@ -99,7 +103,7 @@ class JavaLibPluginKitTest extends AbstractKitTest {
         then: "artifacts deployed"
         deploy.exists()
         def baseName = artifactId + '-1.0'
-        deploy.list() as Set ==
+        withoutModuleFile(deploy) ==
                 ["${baseName}.jar", "${baseName}.pom", "${baseName}-sources.jar", "${baseName}-javadoc.jar", "${baseName}-groovydoc.jar"] as Set<String>
     }
 
@@ -116,7 +120,7 @@ class JavaLibPluginKitTest extends AbstractKitTest {
         """
 
         when: "run pom task"
-        def result = run('install')
+        def result = runVer(GRADLE_VERSION, 'install')
 
 
         String artifactId = projectName()
@@ -129,7 +133,7 @@ class JavaLibPluginKitTest extends AbstractKitTest {
         then: "artifacts deployed"
         deploy.exists()
         def baseName = artifactId + '-1.0'
-        deploy.list() as Set ==
+        withoutModuleFile(deploy) ==
                 ["${baseName}.jar", "${baseName}.pom", "${baseName}-sources.jar"] as Set<String>
     }
 
@@ -148,7 +152,7 @@ class JavaLibPluginKitTest extends AbstractKitTest {
         """
 
         when: "run pom task"
-        def result = run('install')
+        def result = runVer(GRADLE_VERSION, 'install')
 
 
         String artifactId = projectName()
@@ -161,7 +165,59 @@ class JavaLibPluginKitTest extends AbstractKitTest {
         then: "artifacts deployed"
         deploy.exists()
         def baseName = artifactId + '-1.0'
-        deploy.list() as Set ==
+        withoutModuleFile(deploy)  ==
                 ["${baseName}.jar", "${baseName}.pom", "${baseName}-sources.jar"] as Set<String>
+    }
+
+    def "Check jar modification"() {
+
+        setup:
+        file('src/main/java').mkdirs()
+        build """
+            plugins {
+                id 'java'
+                id 'ru.vyarus.java-lib'
+            }
+
+            group 'ru.vyarus'
+            version 1.0
+        """
+
+        when: "run pom task"
+        def result = runVer(GRADLE_VERSION, 'install')
+
+
+        String artifactId = projectName()
+        String baseName = artifactId + '-1.0'
+        ZipFile jar = new ZipFile(file("build/repo/ru/vyarus/$artifactId/1.0/${baseName}.jar"))
+
+        then: "jar manifest correct"
+        String manifest = jar.getInputStream(jar.getEntry('META-INF/MANIFEST.MF')).text
+        println manifest
+        manifest.contains("Implementation-Title: $artifactId")
+        manifest.contains("Implementation-Version: 1.0")
+        manifest.contains("Built-By: ${System.getProperty('user.name')}")
+        manifest.contains("Built-Date:")
+        manifest.contains("Built-JDK:")
+        manifest.contains("Built-Gradle:")
+        manifest.contains("Target-JDK:")
+
+        then: "jar contains pom"
+        String jarPom = jar.getInputStream(jar.getEntry("META-INF/maven/ru.vyarus/$artifactId/pom.xml")).text
+        jarPom != null
+        println jarPom
+
+        then: "jar contains pom.properties"
+        String props = jar.getInputStream(jar.getEntry("META-INF/maven/ru.vyarus/$artifactId/pom.properties")).text
+        props != null
+        println props
+        props.contains('groupId: ru.vyarus')
+        props.contains("artifactId: ${projectName()}")
+        props.contains('version: 1.0')
+    }
+
+    private Set<String> withoutModuleFile(File deployDir) {
+        // gradle 6 publish additional module file (for extended dependencies model)
+        (deployDir.list() as Set<String>).findAll {!it.endsWith('.module')}
     }
 }